Move SPECcurses to mainCLI action, fix imports
