+++ mainCLI.py
@@ -4,6 +4,7 @@ Usage:
     mainCLI.py run [options] <config> [--props <props>]
     mainCLI.py validate [options] <config>
     mainCLI.py dialogue [options]
+    mainCLI.py curses
     mainCLI.py (-h | --help)
     mainCLI.py --version
 
@@ -26,6 +27,7 @@ import dialogue
 from src import validate
 from src import run_generator
 from src import benchmark_run
+from src import speccurses
 
 
 def to_list(s):
@@ -139,6 +141,8 @@ def do_run(arguments):
 
         s.run()
 
+def do_curses(arguments):
+    speccurses.main()
 
 # dictionary of runnables
 # these are functions that take arguments from the
@@ -147,6 +151,7 @@ do = {
     'run': do_run,
     'validate': do_validate,
     'dialogue': do_dialogue,
+    'curses': do_curses,
 }
 
 if __name__ == "__main__":
